+++ luxonis_train/core/__init__.py
@@ -1,8 +1,3 @@
-from .archiver import Archiver
-from .core import Core
-from .exporter import Exporter
-from .inferer import Inferer
-from .trainer import Trainer
-from .tuner import Tuner
+from .core import LuxonisModel
 
-__all__ = ["Exporter", "Trainer", "Tuner", "Inferer", "Archiver", "Core"]
+__all__ = ["LuxonisModel"]
