+++ luxonis_train/models/__init__.py
@@ -1,5 +1,5 @@
-from .luxonis_model import LuxonisModel
+from .luxonis_lightning import LuxonisLightningModule
 from .luxonis_output import LuxonisOutput
 from .predefined_models import *
 
-__all__ = ["LuxonisModel", "LuxonisOutput"]
+__all__ = ["LuxonisLightningModule", "LuxonisOutput"]
