+++ tests/integration/test_sanity.py
@@ -5,7 +5,7 @@ import pytest
 from luxonis_ml.data import LuxonisDataset
 from multi_input_modules import *
 
-from luxonis_train.core import Exporter, Inferer, Trainer, Tuner
+from luxonis_train.core import LuxonisModel
 from luxonis_train.utils.config import Config
 
 TEST_OUTPUT = Path("tests/integration/_test-output")
@@ -20,14 +20,15 @@ OPTS = {
     "trainer.callbacks": "[]",
     "tracker.save_directory": str(TEST_OUTPUT),
     "tuner.n_trials": 4,
+    "exporter.model_name": "test_export",
 }
 
 
 @pytest.fixture(scope="function", autouse=True)
 def clear_output():
     Config.clear_instance()
-    yield
     shutil.rmtree(TEST_OUTPUT, ignore_errors=True)
+    yield
     STUDY_PATH.unlink(missing_ok=True)
     ONNX_PATH.unlink(missing_ok=True)
     shutil.rmtree(INFER_PATH, ignore_errors=True)
@@ -37,34 +38,31 @@ def clear_output():
     "config_file", [str(path) for path in Path("configs").glob("*model*")]
 )
 def test_simple_models(config_file: str):
-    trainer = Trainer(
-        config_file,
-        opts=OPTS,
-    )
-    trainer.train()
-    trainer.test()
-
-    Exporter(config_file).export("test_export.onnx")
+    model = LuxonisModel(config_file, opts=OPTS)
+    model.train()
+    model.test()
+    model.export()
+    assert Path(model.run_save_dir, "export", "test_export.onnx").exists()
 
 
 def test_multi_input():
     config_file = "configs/example_multi_input.yaml"
-    trainer = Trainer(config_file, opts=OPTS)
-    trainer.train()
-    trainer.test(view="val")
+    model = LuxonisModel(config_file, opts=OPTS)
+    model.train()
+    model.test(view="val")
 
     assert not ONNX_PATH.exists()
-    Exporter(config_file).export(str(ONNX_PATH))
+    model.export(str(ONNX_PATH))
     assert ONNX_PATH.exists()
 
     assert not INFER_PATH.exists()
-    Inferer(config_file, view="val", save_dir=INFER_PATH).infer()
+    model.infer(view="val", save_dir=INFER_PATH)
     assert INFER_PATH.exists()
 
 
 def test_custom_tasks(parking_lot_dataset: LuxonisDataset):
     config_file = "tests/configs/parking_lot_config.yaml"
-    Trainer(
+    LuxonisModel(
         config_file,
         opts=OPTS
         | {
@@ -74,6 +72,6 @@ def test_custom_tasks(parking_lot_dataset: LuxonisDataset):
 
 
 def test_tuner():
-    tuner = Tuner("configs/example_tuning.yaml", opts=OPTS)
-    tuner.tune()
+    model = LuxonisModel("configs/example_tuning.yaml", opts=OPTS)
+    model.tune()
     assert STUDY_PATH.exists()
